requirements.txt
# OCR API Service Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pytesseract==0.3.13
Pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
pydantic==2.10.3
pydantic-settings==2.6.1
slowapi==0.1.9
jinja2==3.1.4
python-dotenv==1.0.1
